add the fw refactor
Same structure from the diag and mfg was applied in the firmware function
refactor too. Similar to mfg fw_common.h was created to bridge common
function implementations across generations.
