+++ lib/gen4/fw_gen4.h
@@ -27,4 +27,24 @@
 
 #include "../switchtec_priv.h"
 
+int switchtec_fw_img_write_hdr_gen4(int fd, struct switchtec_fw_image_info *info);
+struct switchtec_fw_part_summary *switchtec_fw_part_summary_gen4(struct switchtec_dev *dev);
+int switchtec_fw_part_info_gen4(struct switchtec_dev *dev, int nr_info, struct switchtec_fw_image_info *info);
+int switchtec_fw_file_info_gen4(int fd, struct switchtec_fw_image_info *info);
+int switchtec_get_device_id_bl2_gen4(struct switchtec_dev *dev, unsigned short *device_id);
+int switchtec_fw_toggle_active_partition_gen4(struct switchtec_dev *dev,
+                                              int toggle_bl2, int toggle_key,
+                                              int toggle_fw, int toggle_cfg,
+                                              int toggle_riotcore);
+int switchtec_fw_write_file_gen4(struct switchtec_dev *dev, FILE *fimg,
+			         int dont_activate, int force,
+			         void (*progress_callback)(int cur, int tot));
+int switchtec_fw_read_gen4(struct switchtec_dev *dev, unsigned long addr, size_t len, void *buf);
+int switchtec_fw_read_fd_gen4(struct switchtec_dev *dev, int fd, unsigned long addr, 
+			      size_t len, void (*progress_callback)(int cur, int tot));
+int switchtec_fw_map_get_active_gen4(struct switchtec_dev *dev, struct switchtec_fw_image_info *info);
+int switchtec_fw_body_read_fd_gen4(struct switchtec_dev *dev, int fd,
+	                           struct switchtec_fw_image_info *info,
+                                   void (*progress_callback)(int cur, int tot));
+
 #endif
