+++ lib/gen6/fw_gen6.h
@@ -27,4 +27,20 @@
 
 #include "../switchtec_priv.h"
 
+int switchtec_fw_img_write_hdr_gen6(int fd, struct switchtec_fw_image_info *info);
+int switchtec_fw_part_info_gen6(struct switchtec_dev *dev, int nr_info,
+				struct switchtec_fw_image_info *info);
+struct switchtec_fw_part_summary *switchtec_fw_part_summary_gen6(struct switchtec_dev *dev);
+struct switchtec_fw_image_info *switchtec_fw_part_data_bl2_gen6(struct switchtec_dev *dev);
+int switchtec_fw_set_redundant_flag_gen6(struct switchtec_dev *dev, int keyman,
+					 int riot, int bl2, int cfg, int fw,
+					 int set);
+int switchtec_fw_img_get_gen6(struct switchtec_dev *dev, int fd,
+			      enum switchtec_fw_type_gen6 fw_type, int fw_slot,
+			      void (*progress_callback)(int cur, int tot));
+int switchtec_fw_toggle_active_partition_gen6(struct switchtec_dev *dev,
+					      int toggle_bl2, int toggle_key,
+					      int toggle_fw, int toggle_cfg,
+					      int toggle_riotcore);
+
 #endif
